-Fixed imcompatibility between shorten links and highlight notes text paterns. -Merged highlight texts claas functions to the TransparentRichTextBox class.
