+++ Master/NucleusCoopTool/UI/UI_Functions.cs
@@ -279,7 +279,6 @@ namespace Nucleus.Coop.UI
                 UI_Interface.HandlerNotesZoom.TextBox.Text = Core_Interface.Current_GenericGameInfo.Description;
                 UI_Interface.HandlerNotesZoom.Visible = true;
                 UI_Interface.HandlerNotesZoom.BringToFront();
-                HighlightNotesText.Highlight(UI_Interface.HandlerNotesZoom.TextBox);
             }
             else
             {
@@ -301,7 +300,6 @@ namespace Nucleus.Coop.UI
                 UI_Interface.HandlerNotesZoom.Notes.Text = Core_Interface.Current_GenericGameInfo.Description;
                 UI_Interface.HandlerNotesZoom.Notes.Rtf = UI_Interface.HandlerNotes.Rtf;
                 UI_Interface.HandlerNotesZoom.Notes.Font = UI_Interface.HandlerNotesZoom.DefaultNotesFont;//Update the font after setting rtf
-                HighlightNotesText.Highlight(UI_Interface.HandlerNotesZoom.TextBox);
             }
             else
             {
@@ -309,7 +307,6 @@ namespace Nucleus.Coop.UI
                 UI_Interface.HandlerNotesZoom.Notes.Rtf = UI_Interface.HandlerNotes.Rtf;
                 UI_Interface.HandlerNotesZoom.Notes.Font = UI_Interface.HandlerNotesZoom.DefaultNotesFont;//Upate the font after setting rtf
                 UI_Interface.HandlerNotesZoom.Warning.Visible = false;
-                HighlightNotesText.Highlight(UI_Interface.HandlerNotesZoom.TextBox);
             }
         }
