+++ Master/NucleusGaming/Controls/TransparentRichTextBox.cs
@@ -1,4 +1,7 @@
-﻿using System.Drawing;
+﻿using System.Collections.Generic;
+using System.Drawing;
+using System.Linq;
+using System.Text.RegularExpressions;
 using System.Windows.Forms;
 
 namespace Nucleus.Gaming.Controls
@@ -26,9 +29,11 @@ namespace Nucleus.Gaming.Controls
         }
 
         private void OnTextChanged(object sender, object e)
-        {       
+        {
             try
             {
+                Highlight(this);
+
                 string shorten = null;
                 int shStart = Text.IndexOf('[');
                 int shEnd = Text.IndexOf(']');
@@ -40,7 +45,7 @@ namespace Nucleus.Gaming.Controls
                 else
                 {
                     Rtf = Rtf.Replace("\\red0\\green0\\blue255;", linkColor); //Go back to blue without this here (unreadable).
-                    ScrollToCaret();//else it won't scroll anymore
+                    ScrollToCaret();//else it won't scroll anymore                   
                     return;
                 }
 
@@ -60,7 +65,7 @@ namespace Nucleus.Gaming.Controls
                     string rtfFormated = FormatRtf(link, shorten);
 
                     if (rtfFormated != Rtf)
-                    {                       
+                    {
                         Rtf = rtfFormated;
                     }
                 }
@@ -79,7 +84,7 @@ namespace Nucleus.Gaming.Controls
                 int slength = Rtf.LastIndexOf("}");
 
                 var sresult = Rtf.Substring(sstart, slength - sstart);
-                var cleanResult = sresult.Remove(sresult.LastIndexOf("}") + 1 );
+                var cleanResult = sresult.Remove(sresult.LastIndexOf("}") + 1);
 
                 var replaceTemplate = "{\\field{\\*\\fldinst HYPERLINK \"" + hyperlink + "\"}{\\fldrslt " + shorten + " }}";
 
@@ -93,18 +98,113 @@ namespace Nucleus.Gaming.Controls
 
                 for (int i = 0; i < diffLenght; i++)//Need to recreate the og lenght or the link won't work when clicked.
                 {
-                     newStr += " ";            
+                    newStr += " ";
                 }
 
                 newRtf = newRtf.Insert(insertIndex, newStr);
 
-                return newRtf.Replace("\\red0\\green0\\blue255;", linkColor); 
+                return newRtf.Replace("\\red0\\green0\\blue255;", linkColor);
             }
             catch
             {
                 return Rtf;
             }
         }
-     
+
+        private static Color defColor = Color.GreenYellow;
+
+        public static void Highlight(RichTextBox textbox)
+        {
+            var text = textbox.Rtf;
+
+            string pattern = @"\*\<([^\[\]\<\>]+)\>\*";
+            var matches = Regex.Matches(text, pattern);
+
+            Dictionary<string, Color> parsedDatas = new Dictionary<string, Color>();
+
+            int index = 0;
+
+            foreach (var match in matches)
+            {
+                var datas = ParseTextDatas(match.ToString());
+
+                if (parsedDatas.Keys.Contains(datas.Item1))
+                {
+                    index++;
+                    continue;
+                }
+
+                parsedDatas.Add(datas.Item1, datas.Item2);
+                textbox.Rtf = textbox.Rtf.Replace(matches[index].ToString(), datas.Item1);
+                index++;
+            }
+
+            foreach (KeyValuePair<string, Color> keyValue in parsedDatas)
+            {
+                int startIndex = 0;
+                while (startIndex < textbox.TextLength)
+                {
+                    int foundIndex = textbox.Find(keyValue.Key, startIndex, RichTextBoxFinds.MatchCase | RichTextBoxFinds.WholeWord);
+                    if (foundIndex == -1) break;
+
+                    textbox.Select(foundIndex, keyValue.Key.Length);
+                    textbox.SelectionColor = keyValue.Value;
+
+                    startIndex = foundIndex + keyValue.Key.Length;
+                }
+            }
+        }
+
+        private static (string, Color) ParseTextDatas(string datas)
+        {
+            string[] splitted = datas.Split('|');
+
+            List<string> cleaned = new List<string>();
+
+            for (int i = 0; i < splitted.Length; i++)
+            {
+                var s = splitted[i];
+                if (s.StartsWith("*<"))
+                {
+                    s = s.Remove(0, 2);
+                }
+                else
+                {
+                    s = s.Remove(s.Length - 2, 2);
+                }
+
+                cleaned.Add(s);
+            }
+
+            if (cleaned.Count == 1)
+            {
+                return (cleaned[0], defColor);
+            }
+            else if (splitted.Length == 2)
+            {
+                return (cleaned[0], ParseColor(cleaned[1]));
+            }
+
+            return ("", defColor);
+        }
+
+        private static Color ParseColor(string colorText)
+        {
+            var colorArray = colorText.Split(',');
+            int r = int.Parse(colorArray[0]); int g = int.Parse(colorArray[1]); int b = int.Parse(colorArray[2]);
+
+            if (colorArray.Length == 3)
+            {
+                return Color.FromArgb(255, r < 0 || r > 255 ? 255 : r,
+                                           g < 0 || g > 255 ? 255 : g,
+                                           b < 0 || b > 255 ? 255 : b);
+            }
+            else
+            {
+                return defColor;
+            }
+        }
+
     }
+
 }
